requirements.txt
PyQt6
pynput
openai
httpx
requests
certifi
anyio
python-dotenv
